requirements.txt
annotated-types   == 0.6.0   
anyio             == 3.7.1   
bcrypt            == 4.1.2   
click             == 8.1.7   
colorama          == 0.4.6   
dnspython         == 2.4.2   
fastapi           == 0.106.0 
h11               == 0.14.0  
httptools         == 0.6.1   
idna              == 3.6     
Jinja2            == 3.1.2   
MarkupSafe        == 2.1.3   
passlib           == 1.7.4   
pip               == 22.3.1  
pydantic          == 2.5.3   
pydantic_core     == 2.14.6  
pymongo           == 4.6.1   
python-dotenv     == 1.0.0   
python-multipart  == 0.0.6   
PyYAML            == 6.0.1   
setuptools        == 65.5.0  
sniffio           == 1.3.0   
starlette         == 0.27.0  
typing_extensions == 4.9.0   
uvicorn           == 0.25.0  
watchfiles        == 0.21.0  
websockets        == 12.0    